Update check should be async
The updater is single flighted so we want the ticker to be async. If an
old update is blocked (a prompt is up and you are away from computer
for a long time), this allows the update check to keep occurring.

This fixes the issue where coyne came into work on Monday and still had
an old update prompt.