App restart (darwin)
This uses ps to find the pid of an app, since app bundles use an executable
like Test.app/Contents/MacOS/Test. It will SIGTERM this process (and SIGKILL
after a second), and then call open.

It doesn't fail to call open if the kill failed.

Bundled is a Test.app that is used in the test.