Use a temp dir to move app before applying update (instead of removing)
Eventhough it's theoretically ok to remove a file executable while it's running,
this is probably safer, especially when there may be underlying OS things like
Gatekeeper doing unknown things.

There could be a scenario in something like an app bundle where if the update applied
and then the app loaded a bundle resource before the restart. This would be
extremely unlikely and probably wouldn't matter since the app would be restarted
but again, probably better to be safe.

This PR creates a temp dir at start of the update which is passed to apply.
At the end of update it is removed (relies on defer).